// test/src/models/enums/languages.enum.ts
export const enum LanguagesEnum {
  ENGLISH = '2057',
  KOREAN = '1042',
  JAPANESE = '1041',
  THAI = '1054',
  RUSSIAN = '1049',
  TURKISH = '1055',
  SPANISH = '1034',
  DANISH = '1030',
  GERMAN = '1031',
  GREEK = '1032',
  FRENCH = '1036',
  HUNGARIAN = '1038',
  ITALIAN = '1040',
  DUTCH = '1043',
  PORTUGUESE = '1046',
  INDONESIAN = '1057',
  MALAY = '1086',
  PERSIAN = '1065',
  JAPAN = '1041',
  CHINESE = '2052',
}

export const languageList = [
  LanguagesEnum.ENGLISH,
  LanguagesEnum.KOREAN,
  LanguagesEnum.JAPANESE,
  LanguagesEnum.THAI,
  LanguagesEnum.RUSSIAN,
  LanguagesEnum.TURKISH,
  LanguagesEnum.SPANISH,
  LanguagesEnum.DANISH,
  LanguagesEnum.GERMAN,
  LanguagesEnum.GREEK,
  LanguagesEnum.FRENCH,
  LanguagesEnum.HUNGARIAN,
  LanguagesEnum.ITALIAN,
  LanguagesEnum.DUTCH,
  LanguagesEnum.PORTUGUESE,
  LanguagesEnum.INDONESIAN,
  LanguagesEnum.MALAY,
  LanguagesEnum.PERSIAN,
];